+++ src/main/org/excelsi/nausicaa/ca/UpdateWeightTransform.java
@@ -5,7 +5,7 @@ import java.util.Random;
 
 
 public class UpdateWeightTransform implements Transform {
-    private static final float MIN_WEIGHT = 0.0001f;
+    private static final double MIN_WEIGHT = 0.0001f;
     private final Random _rand;
 
 
@@ -16,16 +16,16 @@ public class UpdateWeightTransform implements Transform {
     public String name() { return "UpdateWeight"; }
 
     public CA transform(CA c) {
-        float a = (float)_rand.nextGaussian()/8f;
-        float ow = c.getWeight();
-        float nw = Math.max(MIN_WEIGHT,Math.min(1f,a+ow));
+        double a = (double)_rand.nextGaussian()/8f;
+        double ow = c.getWeight();
+        double nw = Math.max(MIN_WEIGHT,Math.min(1d,a+ow));
         System.err.println("old weight: "+ow+", new weight: "+nw+", a: "+a);
         return c.weight(nw);
     }
 
-    public static float mutateWeight(float ow, Random r) {
-        float a = (float)r.nextGaussian()/8f;
-        float nw = Math.max(MIN_WEIGHT,Math.min(1f,a+ow));
+    public static double mutateWeight(double ow, Random r) {
+        double a = (double)r.nextGaussian()/8f;
+        double nw = Math.max(MIN_WEIGHT,Math.min(1d,a+ow));
         System.err.println("old weight: "+ow+", new weight: "+nw+", a: "+a);
         return nw;
     }
